GetUserAPI call works when set to return a model. Working on getting it to work with IAction.
Need to update the access token and call again if it fails.

+++ StreamWorks.Api.Twitch/Models/User/GetUserDataModel.cs
@@ -19,5 +19,5 @@ public class GetUserDataModel
     public string? ProfileImageUrl { get; set; }
     public string? Type { get; set; }
     [JsonPropertyName("view_count")]
-    public int ViewCount { get; set; }
+    public long ViewCount { get; set; }
 }
